refactor: move oscil generators to common for reusability
